chore: move daysBetween function from discount-util to util/dates
