+++ src/utils/dates.test.ts
@@ -1,9 +1,29 @@
 
 import { describe, it, expect } from "vitest";
-import { formatDate, formatDateShort, formatDateLong, formatMonthLong, formatMonthYear, formatYearNumeric } from "./dates";
+import { formatDate, formatDateShort, formatDateLong, formatMonthLong, formatMonthYear, formatYearNumeric, daysBetween } from "./dates";
 
 describe("util", () => {
 
+  describe("daysBetween", () => {
+    it("should calculate days between two dates correctly", () => {
+      expect(daysBetween(new Date(2009, 0, 3), new Date(2009, 0, 3))).toBe(0);
+      expect(daysBetween(new Date(2009, 0, 3), new Date(2009, 0, 4))).toBe(1);
+      expect(daysBetween(new Date(2009, 0, 3), new Date(2009, 0, 2))).toBe(-1);
+      expect(daysBetween(new Date(2009, 0, 3), new Date(2010, 0, 3))).toBe(365);
+      expect(daysBetween(new Date(2024, 11, 6), new Date(2025, 2, 31))).toBe(115);
+  
+      // leap year
+      expect(daysBetween(new Date(2004, 0, 3), new Date(2005, 0, 3))).toBe(366);
+  
+      // with time
+      expect(daysBetween(new Date(2009, 0, 3, 1, 1, 1), new Date(2009, 0, 3, 2, 2, 2))).toBe(0);
+      expect(daysBetween(new Date(2009, 0, 3, 1, 1, 1), new Date(2009, 0, 4, 2, 2, 2))).toBe(1);
+      expect(daysBetween(new Date(2009, 0, 3, 1, 1, 1), new Date(2009, 0, 4, 23, 23, 23))).toBe(1);
+      expect(daysBetween(new Date(2009, 0, 3, 23, 59, 59), new Date(2009, 0, 4, 0, 0, 0))).toBe(1);
+      expect(daysBetween(new Date(2009, 0, 3, 0, 0, 1), new Date(2009, 0, 2, 23, 59, 59))).toBe(-1);
+    });
+  });
+  
   describe("formatDate", () => {
     it("should format date according to designs (default)", () => {
       expect(formatDate(new Date(2009, 0, 3), "en-US")).toBe("03-Jan-09");
